refactor navbar 🎨
