+++ src/components/Badge/Badge.jsx
@@ -1,0 +1,7 @@
+import styles from "./Badge.module.css";
+
+const Badge = ({ children }) => {
+  return <span className={`badge-i ${styles.badgeSmall}`}>{children}</span>;
+};
+
+export default Badge;
