+++ src/components/Badge/Badge.module.css
@@ -1,0 +1,8 @@
+.badgeSmall {
+  right: -0.7rem;
+  top: -0.4rem;
+  font-size: 0.8rem;
+  padding: 0.01rem 0.3rem;
+  background-color: var(--rb-black);
+  border: 0.1rem solid var(--rb-very-light-grey);
+}
