+++ src/components/Navbar/Navbar.module.css
@@ -13,32 +13,33 @@
   justify-content: space-between;
 }
 
-.navBrand {
-  cursor: pointer;
-}
-
 .navRight {
   display: flex;
   align-items: center;
-  justify-content: space-between;
+  justify-content: space-around;
   flex-basis: 60%;
   font-size: 1.3rem;
 }
 
+.navBrand {
+  font-size: 1.6rem !important;
+  font-weight: bold;
+}
+
 .navLinks {
+  font-size: 1.4rem;
   text-decoration: none;
   color: var(--rb-very-light-grey);
 }
 
-.small {
-  right: -0.7rem;
-  top: -0.4rem;
-  font-size: 0.8rem;
-  padding: 0.01rem 0.3rem;
-  background-color: var(--rb-black);
-  border: 0.1rem solid var(--rb-very-light-grey);
+.signinBtn {
+  font-weight: bold;
+  color: var(--rb-black);
+  border: 0.1rem solid var(--rb-grey);
+  background-color: var(--rb-grey) !important;
 }
 
+.signinBtn,
 .signoutBtn {
   padding: 0.2rem 0.5rem;
   margin: 0;
@@ -65,6 +66,10 @@
     flex-basis: 30%;
   }
 
+  .navLinks {
+    font-size: 1.6rem;
+  }
+
   .searchDesktop {
     display: block;
     flex-basis: 40%;
@@ -74,13 +79,19 @@
     padding: 0.5rem;
     width: 100%;
     outline: none;
-    border: 1px solid var(--rb-light-grey);
+    border: 0.1rem solid var(--rb-light-grey);
     background-color: var(--rb-light-grey);
     border-radius: 0.3rem;
   }
 
+  .signinBtn,
   .signoutBtn {
     padding: 0.4rem 0.7rem;
     font-size: 1.1rem;
+  }
+
+  .signinBtn:hover {
+    transition: 0.3s;
+    background-color: inherit !important;
   }
 }
