+++ src/components/Navbar/Search.js
@@ -1,0 +1,17 @@
+import styles from "./Navbar.module.css";
+
+const Search = ({ input, onInput }) => {
+  return (
+    <div className={styles.searchDesktop}>
+      <input
+        value={input}
+        onChange={(e) => onInput(e.target.value)}
+        className="search-box"
+        type="text"
+        placeholder="Search for products"
+      />
+    </div>
+  );
+};
+
+export default Search;
